// src/app/models/userProfileDetails.ts
export interface UserProfileDetails {
  userId?: string;
  registeredEmail?: string;
  firstName: string;
  middleName?: string;
  lastName: string;
  email: string;
  dateOfBirth: string;
  addressLine1: string;
  addressLine2?: string;
  city: string;
  state: string;
  country: string;
  zipcode: string;
  businessName: string;
  businessType: string;
  industry: string;
  domain?: string;
  registrationNumber?: string;
  businessEmail: string;
  businessPhoneNumber: string;
  businessAddressLine1: string;
  businessAddressLine2?: string;
  businessCity: string;
  businessState: string;
  businessCountry: string;
  businessZipcode: string;
  linkedIn?: string;
  facebook?: string;
  instagram?: string;
  twitter?: string;
  youtube?: string;
  businessDescription?: string;
  profilePicture?: string;
  companyLogo?: string;
}